+++ src/constants/config.py
@@ -1,17 +1,19 @@
 import os
 from dotenv import load_dotenv
-from src.utils.utils_date import validate_date_range, format_date
+from src.utils.utils_date import validate_date_range, parse_date
 
 load_dotenv()
+
+ENVIRONMENT_VARIABLE_NAMES = ["START_DATE", "END_DATE", "INTERVAL_IN_DAYS"]
 
 
 def load_environment_variables():
     check_for_environment_variables(
-        ["START_DATE", "END_DATE", "INTERVAL_IN_DAYS"]
+        ENVIRONMENT_VARIABLE_NAMES
     )
 
-    start_date = format_date(os.getenv("START_DATE"))
-    end_date = format_date(os.getenv("END_DATE"))
+    start_date = parse_date(os.getenv("START_DATE"))
+    end_date = parse_date(os.getenv("END_DATE"))
     validate_date_range(start_date, end_date)
 
     return {
